requirements_simple.txt
atlassian-python-api==3.41.10
python-dotenv==1.0.0
google-generativeai==0.8.3
requests==2.31.0